Fix Python3 migration bugs
Several calls to subprocess.Popen were missed when moving to Python3 and
adding universial_newlines = True. This resulted in output being returned
as a binary string instead of the desired string type. A decode was added
to decode the binary strings to strings.

Remove this decode and pass in universal_newlines to the missed
subprocess.Popen calls which was the original intent.